photostudio_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
django-appconf==1.0.6
django-cors-headers==4.4.0
django-fifit lter==24.3
django-imagekit==5.0.0
djangorestframework==3.15.2
importlib_metadata==8.5.0
Markdown==3.7
pilkit==3.0
pillow==10.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
zipp==3.20.2
django-storages==1.14.6
boto3==1.37.38
dj-database-url==3.0.1
psycopg2-binary==2.9.10
